% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_phenotype_prediction_correlation}
\alias{metric_phenotype_prediction_correlation}
\title{Phenotype prediction correlation}
\usage{
metric_phenotype_prediction_correlation(simulation, result)
}
\arguments{
\item{simulation}{A validated benchmark simulation.}

\item{result}{A benchmark result.}
}
\description{
Phenotype prediction correlation
}
