% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_pip_brier}
\alias{metric_pip_brier}
\title{Trait-specific PIP Brier score}
\usage{
metric_pip_brier(simulation, result)
}
\arguments{
\item{simulation}{A validated benchmark simulation.}

\item{result}{A benchmark result.}
}
\description{
Trait-specific PIP Brier score
}
