% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_prediction_correlation}
\alias{metric_prediction_correlation}
\title{Prediction correlation with true genetic value}
\usage{
metric_prediction_correlation(simulation, result)
}
\arguments{
\item{simulation}{A validated benchmark simulation.}

\item{result}{A benchmark result.}
}
\description{
Prediction correlation with true genetic value
}
