% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_prediction_mse}
\alias{metric_prediction_mse}
\title{Prediction MSE against true genetic value}
\usage{
metric_prediction_mse(simulation, result)
}
\arguments{
\item{simulation}{A validated benchmark simulation.}

\item{result}{A benchmark result.}
}
\description{
Prediction MSE against true genetic value
}
