% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_prediction_nmse}
\alias{metric_prediction_nmse}
\title{Normalized prediction MSE against true genetic value}
\usage{
metric_prediction_nmse(simulation, result)
}
\arguments{
\item{simulation}{A validated benchmark simulation.}

\item{result}{A benchmark result.}
}
\description{
Normalized prediction MSE against true genetic value
}
