% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-reporting.R
\name{plot_annotation_alpha_recovery}
\alias{plot_annotation_alpha_recovery}
\title{Plot Study 06 annotation coefficient recovery}
\usage{
plot_annotation_alpha_recovery(alpha)
}
\arguments{
\item{alpha}{Tidy output from \code{annotation_alpha_recovery()}.}
}
\value{
A ggplot object.
}
\description{
Plot Study 06 annotation coefficient recovery
}
