% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-reporting.R
\name{plot_annotation_parameter_recovery}
\alias{plot_annotation_parameter_recovery}
\title{Plot Study 06 variance-parameter recovery}
\usage{
plot_annotation_parameter_recovery(estimates)
}
\arguments{
\item{estimates}{Tidy Study 06 parameter estimates.}
}
\value{
A ggplot object.
}
\description{
Plot Study 06 variance-parameter recovery
}
