% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-reporting.R
\name{plot_benchmark_runtime}
\alias{plot_benchmark_runtime}
\title{Plot benchmark runtime}
\usage{
plot_benchmark_runtime(runtime)
}
\arguments{
\item{runtime}{Tidy runtime rows.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plot benchmark runtime
}
