% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-reporting.R
\name{plot_causal_marker_rank}
\alias{plot_causal_marker_rank}
\title{Plot causal-marker ranks}
\usage{
plot_causal_marker_rank(marker_results)
}
\arguments{
\item{marker_results}{Tidy fine-mapping marker results.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot causal-marker ranks
}
