% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-reporting.R
\name{plot_convergence_mcse}
\alias{plot_convergence_mcse}
\title{Plot Study 04 relative MCSE trajectories}
\usage{
plot_convergence_mcse(convergence, threshold = 0.05)
}
\arguments{
\item{convergence}{Tidy convergence diagnostics.}

\item{threshold}{Prespecified maximum relative MCSE.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot Study 04 relative MCSE trajectories
}
