% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-reporting.R
\name{plot_credible_set_coverage}
\alias{plot_credible_set_coverage}
\title{Plot credible-set coverage}
\usage{
plot_credible_set_coverage(credible_sets)
}
\arguments{
\item{credible_sets}{Tidy credible-set results.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot credible-set coverage
}
