% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-reporting.R
\name{plot_credible_set_size}
\alias{plot_credible_set_size}
\title{Plot credible-set sizes}
\usage{
plot_credible_set_size(credible_sets)
}
\arguments{
\item{credible_sets}{Tidy credible-set results.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot credible-set sizes
}
