% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-reporting.R
\name{plot_operator_spectrum}
\alias{plot_operator_spectrum}
\title{Plot retained spectral mass}
\usage{
plot_operator_spectrum(eigenvalue_summary)
}
\arguments{
\item{eigenvalue_summary}{Block-level retained-eigenvalue table.}
}
\value{
A ggplot object.
}
\description{
Plot retained spectral mass
}
