% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-reporting.R
\name{plot_parameter_bias}
\alias{plot_parameter_bias}
\title{Plot parameter bias}
\usage{
plot_parameter_bias(estimates)
}
\arguments{
\item{estimates}{Tidy parameter-estimate rows.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plot parameter bias
}
