% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-reporting.R
\name{plot_pip_calibration}
\alias{plot_pip_calibration}
\title{Plot posterior-inclusion-probability calibration}
\usage{
plot_pip_calibration(marker_results)
}
\arguments{
\item{marker_results}{Tidy fine-mapping marker results.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot posterior-inclusion-probability calibration
}
