% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{run_sblrbench_method}
\alias{run_sblrbench_method}
\title{Run a benchmark method}
\usage{
run_sblrbench_method(method, ...)
}
\arguments{
\item{method}{A method specification.}

\item{...}{Arguments passed to its fit function.}
}
\description{
Run a benchmark method
}
