% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sblrbench-package.R
\docType{package}
\name{sblrbench-package}
\alias{sblrbench}
\alias{sblrbench-package}
\title{sblrbench: evaluation contracts for sblr methods}
\description{
Strict simulation, alignment, method, result, metric, and provenance
contracts for reproducible benchmarks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/psoerensen/sblrbench}
  \item Report bugs at \url{https://github.com/psoerensen/sblrbench/issues}
}

}
\author{
\strong{Maintainer}: Peter Sørensen \email{peter.sorensen@r-qgg.org}

}
\keyword{internal}
