% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{subset_sblrbench_simulation_samples}
\alias{subset_sblrbench_simulation_samples}
\title{Subset a benchmark simulation to selected samples}
\usage{
subset_sblrbench_simulation_samples(simulation, sample_ids)
}
\arguments{
\item{simulation}{A validated simulation.}

\item{sample_ids}{Requested sample IDs in desired order.}
}
\description{
Subset a benchmark simulation to selected samples
}
