% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{validate_sblrbench_manifest}
\alias{validate_sblrbench_manifest}
\title{Validate a benchmark provenance manifest}
\usage{
validate_sblrbench_manifest(x)
}
\arguments{
\item{x}{Object to validate.}
}
\description{
Validate a benchmark provenance manifest
}
