% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{validate_sblrbench_method}
\alias{validate_sblrbench_method}
\title{Validate a benchmark method specification}
\usage{
validate_sblrbench_method(x)
}
\arguments{
\item{x}{Object to validate.}
}
\description{
Validate a benchmark method specification
}
