% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{validate_sblrbench_result}
\alias{validate_sblrbench_result}
\title{Validate a standard benchmark result}
\usage{
validate_sblrbench_result(x, simulation = NULL)
}
\arguments{
\item{x}{Result to validate.}

\item{simulation}{Optional simulation defining expected identifiers.}
}
\description{
Validate a standard benchmark result
}
