% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{validate_sblrbench_simulation}
\alias{validate_sblrbench_simulation}
\title{Validate a benchmark simulation}
\usage{
validate_sblrbench_simulation(x)
}
\arguments{
\item{x}{Object to validate.}
}
\description{
Validate a benchmark simulation
}
