% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{write_sblrbench_manifest}
\alias{write_sblrbench_manifest}
\title{Write a manifest as pretty JSON}
\usage{
write_sblrbench_manifest(x, path)
}
\arguments{
\item{x}{Manifest.}

\item{path}{Output path.}
}
\description{
Write a manifest as pretty JSON
}
